feat(daemon): Plan 032 G2-core 第七批——get 引用族 9 条 sed 毕业（96→87），tier_router/tracker/provider
auto-lang plan-032-g2-get-borrow 分支（.worktree/plan-032）修复 .get() 返回
Option<&V> 消费链后：key 借用白名单扩枚举/结构体键、is 臂 Some(x) 绑定
return/Ok/let-init 自动 clone、混合臂方法调用 {expr;} 归一。

- retranspile.sh：tier_router 6 + tracker 1 + provider 2 共 9 条 sed 验证
  no-op 后删除（删 sed 重生成与全量 sed 输出逐字节一致）
- tier_router.at：return Some(c) 借用循环变量族经显式 c.clone() 转正
- 三转译 crate 同批重生成（新 auto 含 master 演进）：漂移均为字面量
  as u32 / G2.1 auto-mut 等语义等价形态，随批接受
- 验证：workspace 236 测试 + 转译轨 24 测试全绿 + live e2e 4✅
  （转译版 agent→client→daemon→LLM 全链路，端口 17699 隔离原生实例）

+++ crates/auto-ai-agent/rust/src/agent.rs
@@ -272,7 +272,7 @@ pub struct AgentResult {
 
 impl AgentResult {
     pub fn default() -> AgentResult {
-        return AgentResult { output: "".to_string(), turns: 0, tool_calls: vec![], total_tokens: 0 };
+        return AgentResult { output: "".to_string(), turns: 0 as u32, tool_calls: vec![], total_tokens: 0 as u32 };
     }
     pub fn content(&self) -> String {
         return self.output.clone();
